# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 2.5.0
  description: |
    Valo multi-tenant tender platform API.

    Except for health and organisation discovery/bootstrap operations, protected
    tenant operations resolve an organisation access context on the server. The
    `X-Valo-Organisation-Id` header selects a tenant; it is mandatory when a
    caller has more than one active direct membership, when accessing a client
    through an approved partner relationship, and whenever break-glass access is
    used. The header is only a selector: the server independently validates the
    authenticated user's active membership, role grants, access windows, partner
    relationship, commercial activation, and break-glass approval.
servers:
  - url: /api
    description: Base API path
...
security:
  - clerkBearer: []
paths:
...
  /projects/{id}/evidence:
    get:
      operationId: listEvidence
      tags: [Evidence]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Evidence items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EvidenceItem"
  /projects/{id}/map-evidence:
    post:
      operationId: mapEvidence
      tags: [Evidence]
      summary: Run LLM evidence mapping against confirmed requirements
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Mapping result (suggested evidence items)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceMapResult"
  /evidence:
    post:
      operationId: createEvidence
      tags: [Evidence]
      summary: Manually create an evidence item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvidenceCreate"
      responses:
        "201":
          description: Created evidence item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceItem"
  /evidence/{id}:
    patch:
      operationId: updateEvidence
      tags: [Evidence]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvidenceUpdate"
      responses:
        "200":
          description: Updated evidence item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceItem"
    delete:
      operationId: deleteEvidence
      tags: [Evidence]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    clerkBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk session token. Authentication never implies access to a tenant.
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    EvidenceItem:
      type: object
      required: [id, projectId, requirementId, evidenceStatus, createdAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        requirementId: { type: string }
        requirementText: { type: string, nullable: true }
        documentId: { type: string, nullable: true }
        documentName: { type: string, nullable: true }
        evidenceStatus:
          type: string
          enum: [present, missing, expired, unclear, not_applicable, pending]
        excerpt: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        suggested: { type: boolean }
        confirmedBy: { type: string, nullable: true }
        createdAt: { type: string }
    EvidenceMapResult:
      type: object
      required: [created, items]
      properties:
        created: { type: integer }
        model: { type: string }
        items:
          type: array
          items:
            $ref: "#/components/schemas/EvidenceItem"
    EvidenceCreate:
      type: object
      required: [projectId, requirementId, evidenceStatus]
      properties:
        projectId: { type: string }
        requirementId: { type: string }
        documentId: { type: string }
        evidenceStatus:
          type: string
          enum: [present, missing, expired, unclear, not_applicable, pending]
        excerpt: { type: string }
        notes: { type: string }
    EvidenceUpdate:
      type: object
      properties:
        documentId: { type: string }
        evidenceStatus:
          type: string
          enum: [present, missing, expired, unclear, not_applicable, pending]
        excerpt: { type: string }
        notes: { type: string }
        suggested: { type: boolean }
...